+++ src/main/java/at/jku/oeh/lan/laganizer/model/dao/ArrivalLogDAO.java
@@ -1,6 +1,8 @@
-package at.jku.oeh.lan.laganizer.model.actionlog;
+package at.jku.oeh.lan.laganizer.model.dao;
 
 import org.springframework.data.repository.CrudRepository;
+
+import at.jku.oeh.lan.laganizer.model.ArrivalLog;
 
 import javax.transaction.Transactional;
