+++ src/main/java/at/jku/oeh/lan/laganizer/model/dao/TournamentDAO.java
@@ -1,6 +1,8 @@
-package at.jku.oeh.lan.laganizer.model;
+package at.jku.oeh.lan.laganizer.model.dao;
 
 import org.springframework.data.repository.CrudRepository;
+
+import at.jku.oeh.lan.laganizer.model.Tournament;
 
 public interface TournamentDAO extends CrudRepository<Tournament, Long> {
